ddr/kc.cpp, ddr/kc_m.cpp: Tidied up a little.
